refactor: Name, Price VO 패키지 이동
