+++ src/main/java/kitchenpos/tobe/common/domain/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.tobe.product.domain;
+package kitchenpos.tobe.common.domain;
 
 import java.math.BigDecimal;
 import java.util.Objects;
